% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReticulateMuData.R
\name{as_reticulate_mudata}
\alias{as_reticulate_mudata}
\title{Coerce to ReticulateMuData}
\usage{
as_reticulate_mudata(x)
}
\arguments{
\item{x}{A \link{ReticulateMuData} or a Python \code{mudata.MuData} object.}
}
\value{
A \link{ReticulateMuData}.
}
\description{
If \code{x} is already a \link{ReticulateMuData} it is returned unchanged.  If it is
a Python \code{mudata.MuData} object it is wrapped.  Otherwise an error is raised.
}
\examples{
\dontrun{
mdata2 <- as_reticulate_mudata(py_mdata_obj)
}

}
