% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce_to_anndata.R
\name{.df_to_pandas}
\alias{.df_to_pandas}
\title{Convert an R data.frame to a pandas DataFrame with index = rownames}
\usage{
.df_to_pandas(df, pd)
}
\arguments{
\item{df}{An R \code{data.frame} with row names set.}

\item{pd}{A reticulate handle to the \code{pandas} module.}
}
\value{
A Python pandas \code{DataFrame}.
}
\description{
Convert an R data.frame to a pandas DataFrame with index = rownames
}
\keyword{internal}
