% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReticulateMuData.R
\name{.wrap_py_return}
\alias{.wrap_py_return}
\title{Wrap a Python object in the appropriate R wrapper}
\usage{
.wrap_py_return(obj)
}
\arguments{
\item{obj}{Any R or Python object.}
}
\value{
Wrapped R object or \code{obj} unchanged.
}
\description{
Central dispatch: if \code{obj} is a Python \code{anndata.AnnData} it is wrapped in
\code{anndataR::ReticulateAnnData}; if it is a \code{mudata.MuData} it is wrapped in
\link{ReticulateMuData}.  Any other Python object (or plain R object) is
returned unchanged.
}
\keyword{internal}
