% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReticulateMuData.R
\name{is_reticulate_mudata}
\alias{is_reticulate_mudata}
\title{Test whether an object is a ReticulateMuData}
\usage{
is_reticulate_mudata(x)
}
\arguments{
\item{x}{Any R object.}
}
\value{
\code{TRUE} if \code{x} inherits from \code{"ReticulateMuData"}, \code{FALSE} otherwise.
}
\description{
Test whether an object is a ReticulateMuData
}
\examples{
is_reticulate_mudata(42)        # FALSE

}
