% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReticulateMuData.R
\name{reticulate_mudata}
\alias{reticulate_mudata}
\title{Constructor alias for ReticulateMuData}
\usage{
reticulate_mudata(py_obj)
}
\arguments{
\item{py_obj}{A Python \code{mudata.MuData} object.}
}
\value{
A \link{ReticulateMuData}.
}
\description{
Thin alias for \code{ReticulateMuData$new(py_obj)}.  Equivalent to calling the
constructor directly but reads more naturally in a pipeline.
}
\examples{
\dontrun{
mdata <- reticulate_mudata(py_mdata_from_python)
}

}
