% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basilisk_env.R
\name{with_mofaflex_env}
\alias{with_mofaflex_env}
\title{Evaluate an expression inside the MOFAFlex Python environment}
\usage{
with_mofaflex_env(expr, ..., .packages = "MofaflexR")
}
\arguments{
\item{expr}{An R expression (unquoted) to evaluate.}
}
\value{
The value of \code{expr}.
}
\description{
A thin wrapper around \code{\link[basilisk:basiliskStart]{basilisk::basiliskRun()}} that evaluates \code{expr}
inside the managed Python environment returned by \code{\link[=mofaflex_basilisk_env]{mofaflex_basilisk_env()}}.
Use this wrapper to ensure that all Python imports resolve to the correct
environment.
}
\examples{
\dontrun{
with_mofaflex_env({
  np <- reticulate::import("numpy", convert = FALSE)
  np$array(1:5)
})
}

}
